[-] Basic audio system implemented
Both test SFX and BGM have been added. Can play SFX for now, to check how to implement BGM (should be the same implementation).

How it works (for future me): When a sound is to be played, AudioManager creates an empty GameObject, assign it an AudioSource component, attaches an AudioClip to it and plays sound. After playing, the GameObject is destroyed.

TODO: BGM, Audio Mixer (to be able to control volume levels in menu options)